+++ src/onegov/org/static/cascades/haarcascade_frontalface_default.xml
